pyproject.toml
[tool.poetry]
name = "test-proxy"
version = "0.1.0"
description = ""
authors = ["DAS-SPB <garsiamarkes111@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.11"
pydantic = "^2.10.6"
pyjwt = "^2.10.1"
hvac = "^2.3.0"
regex = "^2024.11.6"
httpx = "^0.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
